requirements.txt
configparser==5.0.0
numpy==1.17.3
timeloop==1.0.2
opencv-python==4.4.0.42
tensorflow==2.3.0
scipy==1.4.1
pyserial==3.4
pyqt5==5.12.0
pyqtwebengine==5.12.0
matplotlib==3.3.1
scikit-learn==0.23.2
psutil==5.7.2
pyqtgraph==0.11.0
Pillow==7.2.0
pyaudio==0.2.11

